Split Wopen3 into separate repository
For some time now Wopen3 has been a dependency of Walrus but built as a
separate gem. It really belongs in its own repository rather than being
bundled up with Walrus.

So this commit moves the Wopen3 files out of the repository and moves all the
stuff that was previously in the "walrus" subdirectory up one level, to the
root of the repository tree.

Signed-off-by: Wincent Colaiuta <win@wincent.com>