Suppress error text for normal exit
The runner calls exit for a couple of normal exit pathways (such as when
supplying the --version or --help switches) and this in turn raises a
SystemExit exception.

We need to explicitly rescue that to prevent an error message from
being emitted to the console (this is because we are now catching
generic exceptions, and SystemExit exceptions are included in that
category).

Signed-off-by: Wincent Colaiuta <win@wincent.com>