Get string_spec.rb working under Ruby 1.9
First of many changes to get Walrus working under Ruby 1.9. Right now we
define a stand-in "jlength" method to replace the one that is no longer
present in 1.9, but eventually will try to make this a little more
elegant but creating new helper methods which call through to the most
appropriate methods depending on which version of Ruby we're running
against.

Signed-off-by: Wincent Colaiuta <win@wincent.com>