Drop jindex extension
As part of the clean-up prior to the extraction of the Walrat gem, bin
the jindex extension.

Note that we drop the "chars" method too, and instead rely on the method
already supplied by Ruby. I believe that this breaks 1.8.6 support, but
the problem can be tackled if people complain about the breakage (I
suspect that nobody on 1.8.6 is using Walrus, so it is unlikely to come
up).

Signed-off-by: Wincent Colaiuta <win@wincent.com>