Set default output extension to empty string
This is an important change because it may break existing workflows.
Previously the default input extension was "tmpl" and the default
output extension was "html".

This meant that given input file:

  hello.tmpl

Compilation would yield:

  hello.rb

And filling would yield:

  hello.html

For these commands:

  walrus compile hello.tmpl
  walrus fill hello.tmpl

This approach isn't ideal because Walrus is agnostic about the language
that will be used for filled templates. It shouldn't make people who
want to target different languages work harder.

So now we maintain the default input extension of "tmpl" but set the
default output extension to "". This means that given input file:

  hello.tmpl

And command line:

  walrus compile hello.tmpl
  walrus fill hello.tmpl

We now get:

  hello.rb (compiled template)
  hello (filled template)

This is the behavior change which may break existing workflows. To get
the old workflow back we instead have to pass the -e or
--output-extension switches to "walrus fill":

  walrus fill -e html hello.tmpl

Or, alternatively, change the source template name:

  hello.html.tmpl

Which will cause:

  walrus compile hello.html.tmpl
  walrus fill hello.html.tmpl

To yield:

  hello.html.rb
  hello.html

One benefit of using multiple extensions like this is that we can now
point Walrus at a folder containing multiple file types; eg:

  display.css.tmpl
  index.html.tmpl

And have it produce the filled output with different extensions:

  display.css
  index.html

Another benefit is that smart text editors can provide better syntax
coloring for files (consider the classic example of a PHP file which
contains HTML text with PHP "islands" embedded inside it).

Signed-off-by: Wincent Colaiuta <win@wincent.com>