Accumulate only last expression in list in an #echo directive
Previously given:

  #echo "foo"; "bar"; "baz"

we echoed "foobarbaz" to the template. Now we evaluate all expressions
in the list but only accumulate the last one (ie. the output is "baz"
although "foo" and "bar" are silently evaluated first).

This means that:

   #echo @foo = 1 + 2; @foo = @foo + 3; @foo

Now prints "6" whereas previously it would have emitted "366".

This brings us into line with most other template languages and also
with Ruby itself:

  #!/usr/bin/env ruby
  puts "result is #{@foo = 1 + 2; @foo = @foo + 3; @foo}"

prints "result is 6", not "result is 366".

Similar examples could be illustrated using Haml and ERB.

Signed-off-by: Wincent Colaiuta <win@wincent.com>