Fix unwanted extra lines after compiled tempalte_body
Previously when there was no "outside" content of a template (ie. no
block definitions) we would end up with compiled output like this:

  def template_body
    ...
  end

  if __FILE__ == $0
    ...
  end

With this commit the unwanted blank line will no longer appear. This
works fine because the DefDirective, which is responsible for producing
the "outside" content, returns it with two trailing newlines when it is
present. So templates with "outside" content continue to look like this:

  def template_body
    ...
  end

  def something_outside
    ...
  end

  if __FILE__ == $0
    ...
  end

Signed-off-by: Wincent Colaiuta <win@wincent.com>