Ruby 1.9: String#each no longer exists
This is probably not the nicest way to fix this under Ruby 1.9, but it
certainly is the easiest way. Given that Walrus is effectively its own
little sandbox doing this kind of meddling with a core class like
String is perhaps a little more justifiable; nevertheless, I'd like to
later on come up with an alternative compatibility fix.

Signed-off-by: Wincent Colaiuta <win@wincent.com>