Add Hextrapolate template files to acceptance spec area
This is a snapshot of the templates that are used to generate the
Hextrapolate application documentation. It is only a snapshot as there
is no need to keep them current with the latest version of the
application; their only purpose is to provide us with a large body of
template files that we can run different versions of Walrus on, on
different Ruby implementations, and confirm that the generated output is
unchanging.

Signed-off-by: Wincent Colaiuta <win@wincent.com>