Fix broken handling of directory arguments
Recursing into directories was broken due to infinite recursion, leading
to "stack level too deep" errors.

On fixing this bug and testing it, I've come to see that the ability to
pass in directories is of limited usefulness anyway, because Walrus
expects all files in the directory to be templates. Seeing as compiled
templates are by default written alongside their inputs, this condition
(of all files in a directory being input templates) will fall apart as
soon as you do a "walrus compile" on the directory.

So I am thinking of removing the "-R" switch and the processing of
directory arguments entirely.

Signed-off-by: Wincent Colaiuta <win@wincent.com>