Work around variable __FILE__ madness
The value of __FILE__ in a spec will vary depending on how it was
invoked, which can wreak havoc with specs that are sensitive to whether
they are working with relative or absolute paths.

The acceptance specs are one example of such sensitive specs, and they
were broken when run as "spec spec" or "spec acceptance", but not as
"spec spec/acceptance/complete_application_spec.rb".

Work around this by first converting to absolute path, and then back to
absolute, so that we know we are always working with relative paths no
matter how we were in invoked.

Signed-off-by: Wincent Colaiuta <win@wincent.com>