Ruby 1.9: Pathname.glob(pathname) chokes
Pathname on 1.9 no longer responds to "to_str", which means that it
can't be passed in to places which expect a string parameter.

Ironically this means that you can't even pass it into all of its own
methods anymore, such as glob:

  Pathname.glob(Pathname.new('.')) # => raises with "can't convert"

This is because under the covers it is delegating to Dir#glob, Dir#glob
complains about the non-string argument.

Signed-off-by: Wincent Colaiuta <win@wincent.com>