Remove special runtime/eval handling from Walrus::Document
As Walrus::Document is an abstract superclass, it will never be
separately run from the command-line, nor read and dynamically
evaluated, so this special handling can be removed.

Actual subclasses, however, can and will use this handling; see the
Walrus::Compiler class which inserts the necessary statements in every
compiled template.

Signed-off-by: Wincent Colaiuta <win@wincent.com>