Gemfile: depend on Git version of Walrat, separate from gemspec
While Bundler's "gemspec" directive is very convenient, it obviously
doesn't allow for the specification of special dependencies like ":path"
or ":git" based gems, because those can't be specified in the gemspec.

Neither can we pull in the gemspec dependencies with "gemspec" and then
redefine them in the Gemfile, because Bundler will complain.

So, the necessary evil here is to repeat our dependencies in both the
gemspec file and the Gemfile. This is the only way to practically
develop a project like Walrus, which has such a tight coupling to
another gem (in this case, Walrat); sometimes changes in Walrat are
required, and it would be insane to have to be able to test them with
Walrus only after publishing a new release of the Walrat gem (because
Bundler tries to force you to use an "isolated" environment, it makes it
very difficult to, for example, install a local copy of the Walrat
development gem to the system, and then reuse that in the Bundle; it
always wants to check rubygems.org to find out what the official latest
release is).

Signed-off-by: Wincent Colaiuta <win@wincent.com>