Unconditionally override String#each
This is ok because we're targeting a specific version of Ruby now, and
will soon switch to using refinements to keep the monkey patching
somewhat clean.

It would be nice to just use `#each_line` directly, but elsewhere we
rely on being able to call `#each` uniformly on objects that might be
arrays or strings. (Duck typing, hmph...)